+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Added
 - **Language**: Added Simplified Chinese language support. (Thanks to @why25!)
 - **Nautilus**: Added "spring-loaded folder" behavior. Dragging an item and hovering over a folder, breadcrumb, tab, or sidebar item for 800ms will automatically open/navigate to it.
-- Readd the Json Editor for rclone fields. (Now it is optional with toggle button. Not like before v0.1.5). This one more user friendly.
+- Readd the Json Editor for rclone fields. (Now it is optional with toggle button. Not like before v0.1.5). This one more user friendly. Now you can add or edit the custom flags from there too.
 - Legacy config directories for docker users.
 - Added `RCLONE_MANAGER_FUSE_COMPAT` environment variable for docker users. This will enable the fuse compatibility mode for rclone.